+++ custom_components/wnhf/executions/generic_real.py
@@ -92,7 +92,7 @@ class GenericExecutionEngine:
     """
 
     API_VERSION = "1.0"
-    VERSION = "1.5-rc3"
+    VERSION = "1.6-rc4"
     ENABLED_ACTIONS = REAL_EXECUTION_ENABLED_ACTIONS
 
     def __init__(